Refactored Pyhon API:
* Introduced flask_restplus -> Swagger
* More scalable API architecture, using flask_restplus namespaces.
* Created a package containing the whole python app. Package name is 'api'.
