+++ api/api/setup.py
@@ -1,0 +1,12 @@
+#!/usr/bin/python
+
+from setuptools import setup
+
+setup(
+    name='api',
+    packages=['api'],
+    include_package_data=True,
+    install_requires=[
+        'flask',
+    ],
+)
