+++ api/run.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python
+
+from api import APP
+APP.run(debug=True)
